Move the right slot of the `MenuSection` outside
It has been done because we had a situation where the one button (one
that contains a title) has another button as its child. This is
forbidden, so now the right slot of this component renders as a sibling.

Additionally, it was necessary to fix some styles in the `MenuLink`
component after all the `MenuSection` changes.